+++ suby/errors.py
@@ -1,2 +1,7 @@
+from suby.subprocess_result import SubprocessResult
+
+
 class RunningCommandError(Exception):
-    pass
+    def __init__(self, message: str, subprocess_result: SubprocessResult) -> None:
+        self.result = subprocess_result
+        super().__init__(self, message)
